Move BasePackageContainer to PackageContainer file
Change BasePackageContainer initializer to internal access
