+++ Sources/PackageGraph/RepositoryPackageContainerProvider.swift
@@ -114,70 +114,6 @@ extension PackageReference {
 
 public typealias RepositoryPackageConstraint = PackageContainerConstraint
 
-/// Base class for the package container.
-public class BasePackageContainer: PackageContainer {
-    public typealias Identifier = PackageReference
-
-    public let identifier: Identifier
-
-    let config: SwiftPMConfig
-
-    /// The manifest loader.
-    let manifestLoader: ManifestLoaderProtocol
-
-    /// The tools version loader.
-    let toolsVersionLoader: ToolsVersionLoaderProtocol
-
-    /// The current tools version in use.
-    let currentToolsVersion: ToolsVersion
-
-    public func versions(filter isIncluded: (Version) -> Bool) -> AnySequence<Version> {
-        fatalError("This should never be called")
-    }
-
-    public var reversedVersions: [Version] {
-        fatalError("This should never be called")
-    }
-
-    public func getDependencies(at version: Version, productFilter: ProductFilter) throws -> [PackageContainerConstraint] {
-        fatalError("This should never be called")
-    }
-
-    public func getDependencies(at revision: String, productFilter: ProductFilter) throws -> [PackageContainerConstraint] {
-        fatalError("This should never be called")
-    }
-
-    public func getUnversionedDependencies(productFilter: ProductFilter) throws -> [PackageContainerConstraint] {
-        fatalError("This should never be called")
-    }
-
-    public func getUpdatedIdentifier(at boundVersion: BoundVersion) throws -> Identifier {
-        fatalError("This should never be called")
-    }
-
-    public func isToolsVersionCompatible(at version: Version) -> Bool {
-        fatalError("This should never be called")
-    }
-
-    fileprivate init(
-        _ identifier: Identifier,
-        config: SwiftPMConfig,
-        manifestLoader: ManifestLoaderProtocol,
-        toolsVersionLoader: ToolsVersionLoaderProtocol,
-        currentToolsVersion: ToolsVersion
-    ) {
-        self.identifier = identifier
-        self.config = config
-        self.manifestLoader = manifestLoader
-        self.toolsVersionLoader = toolsVersionLoader
-        self.currentToolsVersion = currentToolsVersion
-    }
-
-    public var _isRemoteContainer: Bool? {
-        return nil
-    }
-}
-
 /// Local package container.
 ///
 /// This class represent packages that are referenced locally in the file system.
